Fix : next.config
